libadmin-printf-argument-cast-20010623
kill warning by casting argument to printf

+++ src/libadmin/samples/rxstat_get_version.c
@@ -95,7 +95,7 @@ int main(int argc, char *argv[])
     }
 
     printf("\n");
-    printf("RPC stats are version %d\n", version);
+    printf("RPC stats are version %d\n", (int) version);
     printf("\n");
 
     exit(0);
